Refactored PlayerExtendedData into its own files
